% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBoundryArea.R
\name{getBoundaryArea}
\alias{getBoundaryArea}
\title{Get Boundary Area}
\usage{
getBoundaryArea(category, lat, lng)
}
\arguments{
\item{category}{Vector of categories}

\item{lat}{The latitude of the points}

\item{lng}{The longitude of the points}
}
\description{
This function allows you to get the boundary areas of a categorized group of Lat/Long points
}
\keyword{boundary}
