% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmlWriter.r
\name{kmlWriter}
\alias{kmlWriter}
\title{Write KML Boundary File}
\usage{
kmlWriter(g, la, lo, sheetname, filename)
}
\arguments{
\item{g}{Vector of categories / Groups}

\item{la}{The latitude of the points}

\item{lo}{The longitude of the points}

\item{sheetname}{The sheetname specified in the KML File}

\item{filename}{The file name and path for the KML File}
}
\description{
This function allows you to get the boundary areas of a categorized group of Lat/Long points
}
\keyword{boundary}
